+++ src/kernel/network/internet/icmp/ICMP4Message.h
@@ -72,7 +72,7 @@ public:
 
     uint8_t copyDataTo(NetworkByteBlock *byteBlock) override = 0;
 
-    uint16_t getLengthInBytes() override = 0;
+    size_t getLengthInBytes() override = 0;
 
     virtual ICMP4MessageType getICMP4MessageType() = 0;
 };
