+++ src/kernel/network/internet/icmp/messages/ICMP4EchoReply.h
@@ -28,7 +28,7 @@ public:
 
     uint8_t copyDataTo(NetworkByteBlock *byteBlock) override;
 
-    uint16_t getLengthInBytes() override;
+    size_t getLengthInBytes() override;
 
     uint16_t getIdentifier();
