+++ src/kernel/network/internet/icmp/messages/ICMP4TimeExceeded.h
@@ -17,7 +17,7 @@ class ICMP4TimeExceeded : public ICMP4Message {
 
     uint8_t copyDataTo(NetworkByteBlock *byteBlock) override;
 
-    uint16_t getLengthInBytes() override;
+    size_t getLengthInBytes() override;
 
 public:
     ICMP4MessageType getICMP4MessageType() override;
